Sample field documentation.
-- This commit provides an example of what the configuration
documentation might look like.

-- My goal is to make the documentation clear, skimmable, and
searchable, both visually and with tools. The format is as follows:

There are two sections: `Required Fields` and `Optional Fields`. Each
section is a list.

Each entry in the Required Field list has the following format:

```
<li> {@code fieldName} - Type. Brief description of the field's purpose.
```

Each entry in the Optional Field list has the following format:

```
<li> {@code fieldName} - Type. Default: {@code value}. Brief description of the field's purpose.
```

Ideally, the whole thing should fit on one 120 character line, but can
be wrapped if a clear description cannot be condensed into 120ish
characters. If a field is a collection of ObjectNodes, then the fields
on its ObjectNode should be defined in a nested list that follows the
same format.